Remove unnecessary styles
