+++ app/javascript/react/components/MetadataEntry/AcMenuStyles.js
@@ -1,0 +1,15 @@
+const menuStyles = {
+  maxHeight: '180px',
+  overflowY: 'auto',
+  maxWidth: '300px',
+  minWidth: '200px',
+  backgroundColor: '#fafafa',
+  position: 'absolute',
+  top: '1em',
+  zIndex: 1000,
+  padding: 0,
+  listStyle: 'none',
+  // border: '1px solid gray',
+}
+
+export { menuStyles, }
